psql/db-setup/requirements.txt
psycopg2-binary==2.9.6
Faker==19.13.0